requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.11.1
pydantic-settings==2.8.0
python-multipart==0.0.20
httpx==0.28.1
openai==1.58.1
python-dotenv==1.0.1
aiofiles==24.1.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
redis==5.2.1
celery==5.4.0
beautifulsoup4==4.12.3
requests==2.32.3
pandas==2.2.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.3.4
pytest-asyncio==0.24.0
python-docx==1.1.2
PyPDF2==3.0.1
openpyxl==3.1.5
# mcp==1.0.0
python-mcp==1.0.1
selenium==4.27.1
Pillow==11.0.0
websockets==12.0

# Flask web framework for creating REST API
Flask==3.0.0

# Werkzeug for utilities and exceptions handling
Werkzeug==3.0.1

# Additional dependencies that come with Flask
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7

# Development and testing dependencies (optional)
pytest-flask==1.3.0